==> yelp/00-prepare-dataset.ipynb <==
# Prepare the dataset

We load a dataset with messages from customers that are assigned to different categories (invoices, orders, etc.)

We will use this dataset to tune a prompt that can classify a message into one of these categories, with an accuracy as high as possible.

https://huggingface.co/datasets/bitext/Bitext-customer-support-llm-chatbot-training-dataset

pip install 'datasets<3' --quiet

from datasets import load_dataset

dataset = load_dataset("Yelp/yelp_review_full", split="train")
dataset

CLASSES = list(set(dataset["label"]))
CLASSES

### CONSTANTS

N = 200  # Number of samples per class / label
TRAIN_TEST_SPLIT = 0.75  # split dataset in trian and test
TEST_VAL_SPLIT = 0.7  # spit test in test and validation

import pandas as pd
from datasets import Dataset

# Convert the dataset to a pandas DataFrame
df = pd.DataFrame(dataset)
# Add an id column
df = df.reset_index(drop=False).rename(columns={"index": "id"})
# Group by the label column
grouped = df.groupby("label")
# Sample 100 records from each label (adjust N as needed)
sampled_df = grouped.apply(lambda x: x.sample(n=N, random_state=42)).reset_index(
    drop=True
)
# Convert the sampled DataFrame back to a Hugging Face dataset
sampled_dataset = Dataset.from_pandas(sampled_df)
# Print the value counts of the label column
print(sampled_df["label"].value_counts())
# Shuffle the dataset
shuffled_dataset = sampled_dataset.shuffle()

Select only the relevant columns and rename them according to the class ClassificationDataClass

full_dataset = shuffled_dataset.select_columns(["label", "text", "id"])
full_dataset.to_csv("full_dataset.csv")
full_dataset

# Split the dataset into train and test (70% train, 30% test)
train_test_split = full_dataset.train_test_split(test_size=TRAIN_TEST_SPLIT)
# Further split the test set into validation and test sets (50% validation, 50% test)
val_test_split = train_test_split["test"].train_test_split(test_size=TEST_VAL_SPLIT)

train = train_test_split["train"]
train.to_csv("train.csv")
val = val_test_split["train"]
val.to_csv("val.csv")
test = val_test_split["test"]
test.to_csv("test.csv")

train.to_pandas()["label"].value_counts()

val.to_pandas()["label"].value_counts()

test.to_pandas()["label"].value_counts()
