==> yelp/03-fine-tune-bert.ipynb <==
# Classification with BERT

MODEL = "google-bert/bert-base-uncased"
TOKEN_LENGTH = 512

!pip install --upgrade "transformers==4.43.2" accelerate datasets scikit-learn --quiet

import pandas as pd
from datasets import Dataset
from transformers import (
    BertTokenizer,
    BertForSequenceClassification,
    Trainer,
    TrainingArguments,
    pipeline,
)
from sklearn.metrics import accuracy_score, classification_report
import torch

## 1. Read Data

df_train = pd.read_csv("train.csv")
df_test = pd.read_csv("test.csv")
df_val = pd.read_csv("val.csv")

# Convert to Huggingface Dataset
dataset_train = Dataset.from_pandas(df_train)
dataset_val = Dataset.from_pandas(df_val)
dataset_test = Dataset.from_pandas(df_test)

# Load BERT tokenizer
tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")


# Tokenize the dataset
def tokenize_function(examples):
    return tokenizer(
        examples["text"], padding="max_length", truncation=True, max_length=TOKEN_LENGTH
    )


dataset_train = dataset_train.map(tokenize_function, batched=True)
dataset_val = dataset_val.map(tokenize_function, batched=True)

# Set format for PyTorch
dataset_train.set_format(type="torch", columns=["input_ids", "attention_mask", "label"])
dataset_val.set_format(type="torch", columns=["input_ids", "attention_mask", "label"])

## 2. Train Model

# Load pre-trained BERT model
num_classes = df_train["label"].nunique()  # Number of classes in the dataset
model = BertForSequenceClassification.from_pretrained(MODEL, num_labels=num_classes)

# Define TrainingArguments
training_args = TrainingArguments(
    output_dir="./results",
    evaluation_strategy="epoch",
    learning_rate=2e-5,
    per_device_train_batch_size=16,
    per_device_eval_batch_size=16,
    num_train_epochs=3,
    weight_decay=0.01,
    logging_dir="./logs",
    logging_steps=10,
    save_strategy="epoch",
    save_total_limit=2,
    load_best_model_at_end=True,
    metric_for_best_model="accuracy",
)


def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = torch.argmax(torch.tensor(logits), dim=-1)
    return {"accuracy": accuracy_score(labels, predictions)}


trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=dataset_train,
    eval_dataset=dataset_val,
    tokenizer=tokenizer,
    compute_metrics=compute_metrics,
)


# Train the model
trainer.train()
# Evaluate the model
trainer.evaluate()
# Save the model
trainer.save_model("./bert")

## 3. Evaluate

df_test = pd.read_csv("test.csv")

classifier_zero_shot = pipeline(
    "zero-shot-classification", model=MODEL, truncation=True, device=0
)
label_mapping = {
    0: "1 star review",
    1: "2 star review",
    2: "3 star review",
    3: "4 star review",
    4: "5 star review",
}
df_zero_shot = df_test.copy()
df_zero_shot["predictions"] = df_zero_shot["text"].apply(
    lambda x: classifier_zero_shot(x, candidate_labels=list(label_mapping.keys()))[
        "labels"
    ][0]
)

classifier_fine_tuned = pipeline(
    "text-classification", model="./bert", truncation=True, device=0
)
df_fine_tuned = df_test.copy()
df_fine_tuned["predictions"] = df_fine_tuned["text"].apply(
    lambda x: classifier_fine_tuned(x)[0]["label"]
)
df_fine_tuned["predictions"] = (
    df_fine_tuned["predictions"].str.replace("LABEL_", "", regex=False).astype(int)
)

def get_performance_metrics(df_test):
    y_test = df_test.label.round()
    y_pred = df_test.predictions.round()

    print("\nClassification Report:")
    print(classification_report(y_test, y_pred))

# performance using off-the shelf model
get_performance_metrics(df_zero_shot)

# performance of a fine-tuned model
get_performance_metrics(df_fine_tuned)
